Move overlapping output detection to :execution
