+++ subprojects/core/src/test/groovy/org/gradle/api/internal/tasks/execution/DefaultTaskCacheabilityResolverTest.groovy
@@ -24,8 +24,8 @@ import org.gradle.api.internal.tasks.properties.OutputFilePropertySpec
 import org.gradle.api.specs.Spec
 import org.gradle.internal.execution.caching.CachingDisabledReason
 import org.gradle.internal.execution.caching.CachingDisabledReasonCategory
+import org.gradle.internal.execution.history.OverlappingOutputs
 import org.gradle.internal.file.RelativeFilePathResolver
-import org.gradle.internal.fingerprint.overlap.OverlappingOutputs
 import spock.lang.Specification
 
 import javax.annotation.Nullable
